fix 32 vs. 64-bit RNG discrepancies; audit uses of usize with RNG
There were places in the game code where the RNG was being asked to
generate usizes, which are 64 bits native but only 32 bits in wasm32.
This was causing the same seed to produce different results in the
wasm32 build versus the native build.  The wasm32 and native builds
should now produce exactly the same results from now on.

The functional fixes apply to the calculation of the number of items and
monsters to spawn in `spawn::fill_room_with_spawns`, and the selection
weights used in `spawn::spawn_random_item_at`.  The remaining changes
add type annotations to uses of the RNG to ensure that identical types
are used across 32-bit and 64-bit builds, along with a couple of tweaks
to use the RNG API more idiomatically.