make saves persist with IndexedDB in Emscripten build
Saving and returning to the title screen in the Emscripten web build now
preserves the save file across multiple visits using IndexedDB.

Some changes had to be made to make this work.  The "Load Game" option
always shows in the Emscripten build, since the IndexedDB
pseudo-filesystem doesn't initialize in time for the title screen to
detect the presence of the save file at init time.  Saving only persists
when a save actually occurs, so the game will *not* be saved when the
page is closed, for example.

Note: Private browsing windows won't allow saving to IndexedDB, so
saving will quietly fall back to Emscripten's in-memory filesystem and
thus will not persist saves.