add new game plus and win counter
Winning the game now takes the player right back to the start of the
game, resetting the difficulty tracker and incrementing the new win
counter in the process.  The number of wins is shown in the victory
screen if the player has won more than once.

New game plus allows one extra monster and item spawn in a room per win,
limited only by the number of available spaces in the room.  The base
equipment level for spawned weapons and armors is also set upon winning
the game, such that newly spawned equipment becomes increasingly more
powerful as the player continues descending the dungeon.