show hint for pick up item key until an item has been picked up
Stepping on a tile with an item will now append "(Press 'g' to pick
up.)" after the description of the contents of the tile.  This hint is
set to only show when starting a new game and not when loading a saved
game.

The pick up key is the only key that isn't either an immediately obvious
key to try (e.g. arrows, space, enter, escape and to a lesser extent
numpad) or shown on the screen (e.g. "[i] Inventory" in the side bar).
This removes the need for a help screen or for the player to read the
readme to learn the basic controls of the game.