start writing source code guide book
This lays out the book metadata, as well as covering the summary,
introduction and dependency chapters.  The book is compiled into its
published HTML form using mdbook.

The purpose of the guide book is to shed light on the structure and
purpose of the source code of the game in order to make it more useful
as a learning resource for people who are either making their own
roguelike or want to make a roguelike.  It should also satisfy the
curiosity of anybody who wants to know what goes into making a simple
but complete roguelike without the use of roguelike helper libraries.