improve corridor auto-run code with bit masking
This replaces the hard-coded positional tile wall checks and forests of
individual boolean variables with a simpler bit mask approach.

Each tile around the player relative to their movement direction is
stored in a bit within a 16-bit unsigned integer.  This integer is then
masked and matched against desired patterns of walls and open tiles.
The first successful match found returns the direction the player should
move in order to follow the corridor.