change damage formula so high defense reduces weak attacks
When attack falls short of defense doubled, set the final damage to
between 25% to 50% of the attack depending on how far short it falls;
note that attack is typically twice defense, so damage is normally 50%
of the attack value.

This causes higher defense to reduce incoming damage from weak attacks,
so the weaker monsters the player encounters later on are affected by
the player's new defense and armor.