undo chunking and padding of tiles in Map struct
This drastically simplifies size and indexing calculations for map
tiles.  This had originally been done to aid cache friendliness, but I
don't think it's worth the complexity, and it adds a lot of divisions to
critical paths, so out it goes.

This breaks save file compatibility due to rearrangement of saved map
tiles.