show message box on load failure instead of panicking
When picking "Load Game" on the title screen fails to load the save
file, show the error in a message box instead of panicking.  A new
MessageBoxMode has been added to handle this.

This shows "No save file found" when 'savefile.txt' is absent, so this
also gets rid of the gray-out of the "Load Game" option in the
Emscripten build, since it's no longer needed.